+++ fiddles/jquery/fiddle-0038-CorsRequest/app.js
@@ -1,40 +1,26 @@
 (function (app, $, undefined) {
-    "use strict";
-     app.util = function () {
-        return {
-            request: function (url, onSuccess, onError) {
-                $(document).load(url);
-                $(document).ajaxComplete(onSuccess);
-            }
-        }
-    };
-    app.model = function (config) {
-        var localFields = [];
-        return {
-            self: this,
-            fields: localFields
-        }
-    };
-    app.store = function (config) {
-        var localData = [];
-        return {
-            self: this,
-            data: localData,
-            load: function (model) {
-                localData.push(model);
-            }
-        }
-    };
-    app.view = app.view || {
-        render: function (hook) {
-            hook.append('hi there!');
-        },
-        init: function () {
-            var hook = $('#fiddleHook');
-            this.render(hook);
-        }
-    };
-    $(document).ready(function () {
-        app.view.init();
+
+  app.URL = 'http://12-bradyhouse.rhcloud.com/passthru';
+
+  $(document).ready(function () {
+
+    let url = window.location.port ? 'http://' + window.location.hostname + ':' + window.location.port :
+    'http://' + window.location.hostname;
+
+    $('.log').height(window.innerHeight);
+
+
+    $.ajax({
+      type: 'POST',
+      url: app.URL,
+      data: 'url=https://itunes.apple.com/WebObjects/MZStore.woa/wpa/MRSS/featuredalbums/sf=143441/limit=100/genre=20/explicit=true/rss.xml&allowOrigin=' + url + '&allowCredentials=true&allowMethods=GET&allowHeaders=content-type',
+      success: function(data) {
+        $('.log').text(data);
+      },
+      dataType: 'html'
     });
+
+  });
+
+
 })(window.app = window.app || {}, jQuery)