+++ fiddles/node/fiddle-0019-PassThruRest/config/content-types.js
@@ -1,0 +1,17 @@
+module.exports = {
+  'csv'     : 'text/csv',
+  'js'      : 'text/javascript',
+  'ts'      : 'text/typescript',
+  'json'    : 'text/json',
+  'txt'     : 'text/plain',
+  'md'      : 'text/x-markdown',
+  'markdown': 'text/x-markdown',
+  'htm'     : 'text/html',
+  'html'    : 'text/html',
+  'css'     : 'text/css',
+  'tag'     : 'text/html',
+  'ico'     : 'image/x-icon',
+  'jpg'     : 'image/jpeg',
+  'png'     : 'image/png',
+  'svg'     : 'image/svg+xml'
+}